requirements.txt
pandas
streamlit
oracledb
SQLAlchemy
requests
matplotlib
scikit-learn
numpy
openpyxl
seaborn
torch
opencv-python
Pillow
ultralytics